Move precomputed quantities into src
